chore: refactor session-manager stories
