+++ ux/react/src/SessionManager/session-manager.stories.js
@@ -1,0 +1,24 @@
+/* eslint-disable react/jsx-filename-extension */
+
+import { LocaleProvider } from '@arcblock/ux/lib/Locale/context';
+import SessionManager from '.';
+
+export { default as HasLogin } from './demo/has-login';
+export { default as NotLogin } from './demo/not-login';
+export { default as Callback } from './demo/callback';
+export { default as DarkMode } from './demo/dark-mode';
+export { default as Sizes } from './demo/sizes';
+export { default as UsingBlockletService } from './demo/using-blocklet-service';
+
+export default {
+  title: 'SessionManager',
+  component: SessionManager,
+  parameters: { actions: { argTypesRegex: '^on.*' } },
+  decorators: [
+    (Story) => (
+      <LocaleProvider translations={{}}>
+        <Story />
+      </LocaleProvider>
+    ),
+  ],
+};
